+++ Models/Location.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+struct Location: Codable, Hashable {
+    let latitude: Double
+    let longitude: Double
+}